test replace ele

+++ run_pretrain.py
@@ -263,7 +263,7 @@ def model_fn_builder(electra_config, init_checkpoint, learning_rate,
         #whether_replaced = tf.sparse_to_dense(positions, tf.shape(input_ids), diff_cast, default_value=0,
                                               #validate_indices=True, name="whether_replaced")
 
-        zeros = tf.zeros(modeling.get_shape_list(input_ids))
+        zeros = tf.zeros(modeling.get_shape_list(input_ids), dtype=tf.int32)
         whether_replaced = replace_elements_by_indices(zeros, diff, masked_lm_positions)
         print(whether_replaced)
         sys.exit()
